refactor(data): move and refactor OAI-PMH error related exceptions
- For a better overview of exceptions, those directly associated to
  OAI-PMH status codes have been moved into a separate package.
- The exception inheritance hierarchy for these now also reflects better
  their nature: inheriting from an abstract HandlerException that itself
  inherits from xoai-common/OAIException makes these more streamlined.
- The abstract OAIException also requires these exceptions to implement the
  status code method getErrorCode(), which makes them much easier to handle within
  the ErrorHandler later on
