--- xoai-data-provider/src/main/java/io/gdcc/xoai/dataprovider/exceptions/DuplicateDefinitionException.java
@@ -1,20 +1,0 @@
-/*
- * The contents of this file are subject to the license and copyright
- * detailed in the LICENSE and NOTICE files at the root of the source
- * tree and available online at
- *
- * http://www.dspace.org/license/
- */
-package io.gdcc.xoai.dataprovider.exceptions;
-
-/**
- * @author João Melo
- */
-public class DuplicateDefinitionException extends HandlerException {
-    private static final long serialVersionUID = -6874140130853606687L;
-
-    public DuplicateDefinitionException(String message) {
-        super(message);
-    }
-
-}
