+++ xoai-data-provider/src/main/java/io/gdcc/xoai/dataprovider/exceptions/handler/package-info.java
@@ -1,0 +1,1 @@
+package io.gdcc.xoai.dataprovider.exceptions.handler;
